remove python-instagram dependency; call API directly.
some instagram-specific changes to interpret_http_exception:

when error body contains specific types that mean the access token was
revoked or the account was deleted, change the error code to 401 to
tell Bridgy that the source should be disabled.

note that there was a small regression here when we stopped using the
python-instagram library. instagram errors previously were
reformatted as: "error_type: error_message". this version will
simply give back the unformatted json blob, so instead of

  OAuthAccessTokenException: The access_token provided is invalid.

we get:

  {"meta":{"error_type":"OAuthAccessTokenException","code":400,
   "error_message":"The access_token provided is invalid."}}