Migrate IndieAuth to full authorization_code grant
A number of IndieAuth servers do not support the IndieAuth-only
[Profile URL Response], so to widen our support for IndieAuth servers we
can migrate to a full authorization_code token grant.

This allows us to fall back to the existing functionality if no
`token_endpoint` is found, to allow for legacy clients' usage, and for
servers that don't have a token endpoint.

[Profile URL Response]: https://indieauth.spec.indieweb.org/#profile-url-response